requirements.txt
numpy==1.21.4
opencv_python==3.4.16.59
Pillow==8.4.0
pyspark==3.2.0
scikit_image==0.18.3
skimage==0.0
